Conhecendo e implementando um front controller
Aqui ainda não estamos dentro do padrão MVC nem nas boas práticas de
programação, mas o próximo passo é transformar essa aplicação procedural
em uma aplicação usando todos os conceitos de POO.

+++ remover-video.php
@@ -6,13 +6,14 @@ $password = '';
 
 $pdo = new PDO("mysql:host=$host;dbname=$dbname", $username, $password);
 
-$id = $_GET['id'];
+$id = filter_input(INPUT_GET, 'id', FILTER_VALIDATE_INT);
+if ($id !== false && $id !== null)
 $sqlQuery = 'DELETE FROM videos WHERE id = ?';
 $stmt = $pdo->prepare($sqlQuery);
 $stmt->bindValue(1, $id, PDO::PARAM_INT);
 
 if ($stmt->execute() === false) {
-    header('Location:./index.php?sucesso=0' );
+    header('Location:./?sucesso=0' );
 }else {
-    header('Location:./index.php?sucesso=1' );
+    header('Location:./?sucesso=1' );
 }
